Move all session config command implementations from session to device
Previous implementation stretched some way to be able to report
invalid session ids, this format makes it easier to read and
understand each command implementation individually.
